Keep in one component to keep diff small, add transition on positioning
