--- src/shared-components/immersiveModal/constants.ts
@@ -1,3 +1,0 @@
-export const MODAL_MOBILE_SCROLLING_ID = 'modal-mobile-scrolling-id';
-export const MODAL_DESKTOP_SCROLLING_ID = 'modal-desktop-scrolling-id';
-export const CROSS_ICON_CONTAINER_ID = 'cross-icon-container';
