+++ src/shared-components/immersiveModal/index.tsx
@@ -3,17 +3,13 @@ import React, { useEffect, useRef, useState } from 'react';
 import ReactDOM from 'react-dom';
 import { Transition } from 'react-transition-group';
 import throttle from 'lodash.throttle';
-import { FocusScope } from '@react-aria/focus';
+import { FocusScope, useFocusManager } from '@react-aria/focus';
 
 import { REACT_PORTAL_SECTION_ID } from '../../constants/portals';
 import { OffClickWrapper } from '../offClickWrapper';
 import { CrossIcon } from '../../icons';
 import Style from './style';
-import {
-  MODAL_DESKTOP_SCROLLING_ID,
-  MODAL_MOBILE_SCROLLING_ID,
-} from './constants';
-import { ImmersiveModalContent } from './immersiveModalContent';
+import { isDefined } from '../../utils/isDefined';
 
 export interface ImmersiveModalProps {
   /**
@@ -43,6 +39,9 @@ interface ImmersiveModal extends React.FC<ImmersiveModalProps> {
   Paragraph: typeof Style.Paragraph;
 }
 
+const MODAL_MOBILE_SCROLLING_ID = 'modal-mobile-scrolling-id';
+const MODAL_DESKTOP_SCROLLING_ID = 'modal-desktop-scrolling-id';
+
 const getHtmlNode = () => document.querySelector('html') ?? document.body;
 const getDomNode = () =>
   document.getElementById(REACT_PORTAL_SECTION_ID) ?? document.body;
@@ -50,6 +49,97 @@ const getModalMobileScrollingElement = () =>
   document.getElementById(MODAL_MOBILE_SCROLLING_ID);
 const getModalDesktopScrollingElement = () =>
   document.getElementById(MODAL_DESKTOP_SCROLLING_ID);
+
+interface ImmersiveModalContentProps
+  extends Pick<
+    ImmersiveModalProps,
+    'children' | 'footerContent' | 'headerImage' | 'title'
+  > {
+  handleCloseIntent: () => void;
+  showDesktopHeaderBar: boolean;
+  showMobileHeaderBar: boolean;
+}
+
+/**
+ * We define a standalone component, ImmersiveModalContent, in order to add
+ * `useFocusManager` functionality, which only works as a child of <FocusScope>
+ */
+const ImmersiveModalContent = ({
+  footerContent,
+  headerImage,
+  showDesktopHeaderBar,
+  showMobileHeaderBar,
+  title,
+  handleCloseIntent,
+  children,
+}: ImmersiveModalContentProps) => {
+  const focusManager = useFocusManager();
+
+  /**
+   * It is not typical modal behavior to be able to scroll. Consequently, our
+   * keyboard autofocus handling means when users try and scroll in the ImmersiveModal,
+   * nothing happens. We make the content of the container tabbable such that when the
+   * arrow keys are used (as in normal scroll behavior) we are able to break focus with
+   * the close button and allow scrolling of the content
+   */
+  const onKeyDown = (event: React.KeyboardEvent) => {
+    switch (event.key) {
+      case 'ArrowDown':
+        focusManager.focusNext({ wrap: true });
+        break;
+      case 'ArrowUp':
+        focusManager.focusPrevious({ wrap: true });
+        break;
+      default:
+        break;
+    }
+  };
+
+  const hasHeaderImage = isDefined(headerImage);
+
+  return (
+    <Style.MainModalContentContainer
+      id={MODAL_DESKTOP_SCROLLING_ID}
+      hasHeaderImage={hasHeaderImage}
+    >
+      {!showMobileHeaderBar && (
+        <React.Fragment>
+          <Style.DesktopHeaderBar showDesktopHeaderBar={showDesktopHeaderBar}>
+            <span>{title}</span>
+          </Style.DesktopHeaderBar>
+          <Style.CrossIconButton
+            onClick={handleCloseIntent}
+            showDesktopHeaderBar={showDesktopHeaderBar}
+            onKeyDown={onKeyDown}
+          >
+            <CrossIcon />
+          </Style.CrossIconButton>
+        </React.Fragment>
+      )}
+
+      {hasHeaderImage && (
+        <Style.HeaderImageContainer>{headerImage}</Style.HeaderImageContainer>
+      )}
+      {/**
+       * Note: we normally do not want to add tabIndex to non-interactive elements,
+       * but it is necessary to make ImmersiveModal fully keyboard navigable when
+       * there  is scrollable content
+       */}
+      <Style.ContentWithFooterContainer
+        hasHeaderImage={hasHeaderImage}
+        tabIndex={0}
+      >
+        <div>
+          {isDefined(title) && <Style.ModalTitle>{title}</Style.ModalTitle>}
+          <section>{children}</section>
+        </div>
+        {isDefined(footerContent) && (
+          <Style.ModalFooter>{footerContent}</Style.ModalFooter>
+        )}
+      </Style.ContentWithFooterContainer>
+    </Style.MainModalContentContainer>
+  );
+};
 
 /**
  * It is used to provide a layer on top of a page when we need to present more content and actions to patients.
